refactor: split AuthorizationTests by controller
Separated the monolithic AuthorizationTests class into dedicated test classes for each controller (Users, Teams, Sprints, MoodEntries, Auth). This improves test maintainability and readability without changing the test logic.
